Apply pull request suggestions
